💝 Refactor development clusters
Fix the repository structure so that the cluster setup and deployment
configs become importable via regular Go imports.

Downstream consumers and our internal repos can now depend on these two
modules and copy devtools/main.go to quickly set up a development cluster:

github.com/eomii/rules_ll/devtools/deployments
github.com/eomii/rules_ll/devtools/clusters
